refactor(account-detail): Reworked the chest count per reward query
