+++ apps/website/src/actions/chest.ts
@@ -138,67 +138,3 @@ export const getLatestChest = async (accountId?: number) => {
 
   return chest;
 };
-
-export const getMostReceivedReward = async (accountId?: number) => {
-  let query = db
-    .selectFrom("chest")
-    .innerJoin("reward", "chest.rewardId", "reward.id");
-
-  if (accountId) {
-    query = query.where("chest.accountId", "=", accountId);
-  }
-
-  const reward = await query
-    .select((eb) => [
-      "reward.name",
-      eb.fn.count<number>("chest.id").as("count"),
-    ])
-    .groupBy(["reward.name"])
-    .orderBy("count", "desc")
-    .executeTakeFirstOrThrow();
-
-  return reward;
-};
-
-export const getChestCountPerCategory = async (accountId?: number) => {
-  let query = db
-    .selectFrom("chest")
-    .rightJoin("reward", "chest.rewardId", "reward.id")
-    .rightJoin("category", "reward.category", "category.id"); // Include all categories
-
-  if (accountId) {
-    query = query.where("chest.accountId", "=", accountId);
-  }
-
-  const categories = await query
-    .select((eb) => [
-      "category.name",
-      eb.fn.count<number>("chest.id").as("count"),
-    ])
-    .groupBy("category.name")
-    .orderBy("count", "desc")
-    .execute();
-
-  return categories;
-};
-
-export const getChestCountPerRarity = async (accountId?: number) => {
-  let query = db
-    .selectFrom("chest")
-    .innerJoin("rarity", "chest.rarityId", "rarity.id");
-
-  if (accountId) {
-    query = query.where("chest.accountId", "=", accountId);
-  }
-
-  const rarities = await query
-    .select((eb) => [
-      "rarity.name",
-      eb.fn.count<number>("chest.id").as("count"),
-    ])
-    .groupBy(["rarity.name", "rarity.chance"])
-    .orderBy("rarity.chance", "desc")
-    .execute();
-
-  return rarities;
-};
