+++ apps/website/src/actions/event.ts
@@ -98,22 +98,6 @@ export const getAllEvents = async ({
   };
 };
 
-export const getHighestEvent = async () => {
-  const result = await db
-    .selectFrom("event")
-    .leftJoin("chest", "chest.eventId", "event.id")
-    .select((eb) => [
-      "event.name",
-      "event.endDate",
-      eb.fn.count<number>("chest.id").as("count"),
-    ])
-    .groupBy(["event.name", "event.endDate"])
-    .orderBy("count", "desc")
-    .executeTakeFirstOrThrow();
-
-  return result;
-};
-
 export const deriveEventStatus = (
   startDate: Expression<Date>,
   endDate: Expression<Date>,
